Restructuring: Many things have been changed at once.
This type of commit should not occur again.

Changes made to the project:
The project's name became Custom Helper. It makes more sense with what we're trying to do with the software.

The "assets" folder was created for the CSS files, images and some JS used in the entire project.

The JSON files in the OURONET section are now in the "helpers" folder for better organization.

The app will not open more than one instance.

For Devs: When executing the command "npm debug" the app will open in full screen mode with the developer tools.

The menu template has been translated to en-us.

Features:
It is now possible to backup the connections to reuse in the future if the configuration files are deleted (after reinstalling the OuroNet application).

Also included is the option to select a directory other than the default where the configuration files are located.

Some descriptions have been added in the section's process header. The pattern will be followed for all other processes added in the future.